Remove console middleware
