+++ src/devtools/client/webconsole/constants.js
@@ -11,7 +11,7 @@ const actionTypes = {
   AUTOCOMPLETE_PENDING_REQUEST: "AUTOCOMPLETE_PENDING_REQUEST",
   AUTOCOMPLETE_RETRIEVE_FROM_CACHE: "AUTOCOMPLETE_RETRIEVE_FROM_CACHE",
   AUTOCOMPLETE_TOGGLE: "AUTOCOMPLETE_TOGGLE",
-  BATCH_ACTIONS: "BATCH_ACTIONS",
+
   CLEAR_HISTORY: "CLEAR_HISTORY",
   EDITOR_TOGGLE: "EDITOR_TOGGLE",
   EDITOR_ONBOARDING_DISMISS: "EDITOR_ONBOARDING_DISMISS",
